requirements.txt
APScheduler==3.6.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
soupsieve==1.9
SQLAlchemy==1.3.1
tzlocal==1.5.1
urllib3==1.24.1
